Module Tezos_alpha_test_helpers

module Account : sig ... end

module Account_helpers : sig ... end

module Adaptive_issuance_helpers : sig ... end

module Assert : sig ... end

module Big_map_helpers : sig ... end

module Block : sig ... end

module Consensus_helpers : sig ... end

module Constants_helpers : sig ... end

module Context : sig ... end

module Contract_helpers : sig ... end

module Cpmm_logic : sig ... end

module Cpmm_repr : sig ... end

module Dal_helpers : sig ... end

module Dummy_zk_rollup : sig ... end

module Error_helpers : sig ... end

module Expr : sig ... end

module Expr_common : sig ... end

module Incremental : sig ... end

module Liquidity_baking_generator : sig ... end

This module provides a set of abstractions to reason about the so-called “liquidity baking” feature1.

module Liquidity_baking_machine : sig ... end

module Log_helpers : sig ... end

module Lqt_fa12_repr : sig ... end

module Lwt_result_wrap_syntax : sig ... end

Extends the Lwt_result_syntax with additional utilities for wrapping results produced by the protocol, i.e. Environment.Error_monad.tzresult, to tzresult Lwt.t values used in the tests.

module Merkle_list_helper : sig ... end

module Nonce : sig ... end

module Op : sig ... end

module Operation_generator : sig ... end

module Result_wrap_syntax : sig ... end

Extends the Result_syntax with additional utilities for wrapping results produced by the protocol, i.e. Environment.Error_monad.tzresult, to tzresult values used in the tests.

module Rewards : sig ... end

The tables are precomputed using this the following formulas:

module Sapling_helpers : sig ... end

module Sc_rollup_helpers : sig ... end

module Scenario : sig ... end

module Scenario_bake : sig ... end

module Scenario_base : sig ... end

module Scenario_begin : sig ... end

module Scenario_constants : sig ... end

module Scenario_dsl : sig ... end

module Scenario_op : sig ... end

module Script_big_map : sig ... end

module Script_map : sig ... end

module Script_set : sig ... end

module Slashing_helpers : sig ... end

module State : sig ... end

module State_account : sig ... end

module Test_global_constants : sig ... end

module Testable : sig ... end

module Tez_helpers : sig ... end

module Tez_staking_helpers : sig ... end

module Tezt_helpers : sig ... end

module Ticket_helpers : sig ... end

module Transfers : sig ... end

module Zk_rollup_l2_helpers : sig ... end