Module Tezos_alpha_test_helpers.Contract_helpers

val init : ?hard_gas_limit_per_block: Tezos_protocol_alpha.Protocol.Alpha_context.Gas.Arith.integral -> unit -> (Block.t * Tezos_protocol_environment_alpha.Signature.Public_key_hash.t * Tezos_protocol_alpha.Protocol.Alpha_context.Contract.t * Tezos_protocol_alpha.Protocol.Alpha_context.Contract.t, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t

Initializes 2 addresses to do only operations plus one that will be used to bake.

val read_file : string -> string

Return contents of a given file as string.

val load_script : storage:string -> string -> Tezos_protocol_alpha.Protocol.Alpha_context.Script.t

Loads a script from file.

val originate_contract_hash : string -> string -> Tezos_protocol_alpha.Protocol.Alpha_context.Contract.t -> Block.t -> Tezos_protocol_alpha.Protocol.Alpha_context.public_key_hash -> (Tezos_protocol_alpha.Protocol.Contract_hash.t * Block.t, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t

Returns a block in which the contract is originated.

val originate_contract : string -> string -> Tezos_protocol_alpha.Protocol.Alpha_context.Contract.t -> Block.t -> Tezos_protocol_alpha.Protocol.Alpha_context.public_key_hash -> (Tezos_protocol_alpha.Protocol.Alpha_context.Contract.t * Block.t, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t

val fake_KT1 : Tezos_protocol_alpha.Protocol.Contract_hash.t

val default_self : Tezos_protocol_alpha.Protocol.Contract_hash.t

val default_payer : Tezos_base.TzPervasives.Signature.Public_key_hash.t

val default_sender : Tezos_protocol_alpha.Protocol.Alpha_context.Contract.t

val default_step_constants : Tezos_protocol_alpha.Protocol.Script_interpreter.step_constants

val run_script : Tezos_raw_protocol_alpha.Alpha_context.t -> ?logger:Tezos_raw_protocol_alpha.Script_typed_ir.logger -> ?step_constants: Tezos_protocol_alpha.Protocol.Script_interpreter.step_constants -> ?internal:bool -> string -> ?entrypoint:Tezos_protocol_alpha.Protocol.Alpha_context.Entrypoint.t -> storage:string -> parameter:string -> unit -> (Tezos_protocol_alpha.Protocol.Script_interpreter.execution_result * Tezos_raw_protocol_alpha.Alpha_context.context) Tezos_base.TzPervasives.tzresult Lwt.t

Helper function that parses and typechecks a script, its initial storage and parameters from strings. It then executes the typed script with the storage and parameters and returns the result.

The step_constants argument passes in some data which remains constant throughout script's execution, hence the name. This includes addresses of the sender and payer, the address of the smart contract, the amount of Tez transferred to it and so on.

An internal operation is an operation generated by smart contract's execution rather than by an implicit account.

val originate_contract_from_string_hash : script:string -> storage:string -> source_contract:Tezos_protocol_alpha.Protocol.Alpha_context.Contract.t -> baker:Tezos_protocol_alpha.Protocol.Alpha_context.public_key_hash -> Block.t -> (Tezos_protocol_alpha.Protocol.Contract_hash.t * Tezos_protocol_alpha.Protocol.Alpha_context.Script.t * Block.t, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t

val originate_contract_from_string : script:string -> storage:string -> source_contract:Tezos_protocol_alpha.Protocol.Alpha_context.Contract.t -> baker:Tezos_protocol_alpha.Protocol.Alpha_context.public_key_hash -> Block.t -> (Tezos_protocol_alpha.Protocol.Alpha_context.Contract.t * Tezos_protocol_alpha.Protocol.Alpha_context.Script.t * Block.t, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t