Module Dummy_zk_rollup.Operator

Basic rollup operator for generating Updates.

Parameters

module Params : sig ... end

Signature

val init_state : Tezos_protocol_alpha.Protocol.Alpha_context.Zk_rollup.State.t

Initial state of the rollup

val circuits : [ `Public | `Private | `Fee ] Kzg.SMap.t

Map associating every circuit identifier to its kind

val lazy_pp : (Plonk.Main_protocol.prover_public_parameters * Plonk.Main_protocol.verifier_public_parameters) lazy_t

Commitment to the circuits

val craft_update : Tezos_protocol_alpha.Protocol.Alpha_context.Zk_rollup.State.t -> zk_rollup:Tezos_protocol_alpha.Protocol.Alpha_context.Zk_rollup.t -> ?private_ops: Tezos_protocol_alpha.Protocol.Alpha_context.Zk_rollup.Operation.t list list -> ?exit_validities:bool list -> Tezos_protocol_alpha.Protocol.Alpha_context.Zk_rollup.Operation.t list -> Tezos_protocol_alpha.Protocol.Alpha_context.Zk_rollup.State.t * Tezos_protocol_alpha.Protocol.Alpha_context.Zk_rollup.Update.t

craft_update state ~zk_rollup ?private_ops ?exit_validities public_ops will apply first the public_ops, then the private_ops. While doing so, the public inputs for every circuit will be collected. A Plonk proof of correctness of the application these operations is created.

module Internal_for_tests : sig ... end