Module Tezos_alpha_test_helpers.Scenario_begin

type Tezos_base.TzPervasives.error +=

| Inconsistent_number_of_bootstrap_accounts

Returns when the number of bootstrap accounts created by Context.init_n n is not equal to n

type starter_constants =

| Mainnet

| Sandbox

| Test

val start : constants:starter_constants -> (unit, Scenario_constants.constants) Scenario_dsl.scenarios

val start_with : constants:Scenario_constants.constants -> (unit, Scenario_constants.constants) Scenario_dsl.scenarios

val start_with_list : constants:(string * Scenario_constants.constants) list -> (unit, Scenario_constants.constants) Scenario_dsl.scenarios

val init_constants : ?default:starter_constants -> ?reward_per_block:int64 -> ?deactivate_dynamic:bool -> ?blocks_per_cycle:int32 -> ?delegate_parameters_activation_delay:int -> unit -> (unit, Scenario_constants.constants) Scenario_dsl.scenarios

Initializes the constants for testing, with well chosen default values. Recommended over start or start_with

val begin_test : ?burn_rewards:bool -> ?force_attest_all:bool -> Tezos_base.TzPervasives.String.Map.key list -> (Scenario_constants.constants, Scenario_base.t) Scenario_dsl.scenarios

Initialize the test, given some initial parameters