Module Slashing_helpers.Full_denunciation

Helpers about "full denunciations", that is, a denunciation item and its culprit. See type t of this module.

type t = Tezos_base.TzPervasives.Signature.Public_key_hash.t * Tezos_protocol_alpha.Protocol.Denunciations_repr.item

A denunciation item preceded by the culprit's pkh. Indeed, the culprit isn't recorded inside the Protocol.Denunciations_repr.item because it serves as a key in the protocol's storage instead. But we often need both together in the tests.

val check_same_lists_any_order : loc:string -> t list -> t list -> unit Tezos_base.TzPervasives.tzresult Lwt.t

Asserts that both lists contain the same elements.

These elements may be ordered differently, but must have the same multiplicity in both lists.

val slashing_percentage : block_before_slash:Block.t -> Tezos_protocol_alpha.Protocol.Misbehaviour_repr.t -> t list -> Tezos_protocol_alpha.Protocol.Percentage.t Tezos_base.TzPervasives.tzresult Lwt.t

Same as slashing_percentage (top-level function), but takes directly a list of full denunciations as an argument.

The list of full denunciation must contain at least all the denunciations for misbehaviour known to the chain. Note that it may additionally contain denunciations for other misbehaviours, which will be ignored.

See slashing_percentage for the precondition on block_before_slash.