Module Tezos_alpha_test_helpers.State_ai_flags

This module handles the logic of flags for AI/AS in the State

module AI_Activation : sig ... end

This module takes care of the flags force_activation and activation_vote_enable, and updates and check the record field ai_activation_cycle in the state.

module AI : sig ... end

This module only checks (for now) if AI is activated or not

module Delayed_slashing : sig ... end

This module takes care of choosing the denunciations that need to be applied at the end of a cycle. It depends on the flag ns_enable.

module NS : sig ... end

This module takes care of the new adaptive slashing mechanism.