Module State_ai_flags.Delayed_slashing

This module takes care of choosing the denunciations that need to be applied at the end of a cycle. It depends on the flag ns_enable.

val enabled : State.t -> bool

Delayed_slashing is enabled iff ns_enable = true.

val partition_slashes : State.t -> Tezos_protocol_alpha.Protocol.Alpha_context.Cycle.t -> (Tezos_base.TzPervasives.Signature.Public_key_hash.t * Tezos_protocol_alpha.Protocol.Denunciations_repr.item) list * (Tezos_base.TzPervasives.Signature.Public_key_hash.t * Tezos_protocol_alpha.Protocol.Denunciations_repr.item) list

partition_slashes s cycle returns a pair (l1,l2) of lists of slashes, partitioned from the state.pending_slashes. l2 is the list of slashes to apply at the end of the given cycle, and l1 is the rest (which should usually replace state.pending_slashes)