Module State_ai_flags.NS

This module takes care of the new adaptive slashing mechanism.

val enabled : Block.t -> State.t -> bool

It is enabled iff the flag ns_enable is set to true, and AI is also enabled.

val get_double_attestation_slashing_percentage : (Tezos_base.TzPervasives.Signature.public_key_hash * Tezos_protocol_alpha.Protocol.Denunciations_repr.item) list -> Block.t -> State.t -> Tezos_protocol_alpha.Protocol.Misbehaviour_repr.t -> Tezos_protocol_alpha.Protocol.Percentage.t Tezos_base.TzPervasives.tzresult Lwt.t

Whatever the value of the flag is, this function returns the slashing value for a given double attestation