Module Tez_helpers.Ez_tez

Functions to manipulate Tez in a high level way

type tez_quantity =

| Half

| All

| All_but_one

| Nothing

| Max_tez

| Amount of t

Aliases for tez values

val tez_quantity_pp : Stdlib.Format.formatter -> tez_quantity -> unit

val quantity_to_tez : t -> tez_quantity -> t

quantity_to_tez max qty returns a tez value corresponding to the given qty. If qty is All, then returns max. If qty is All_but_one, returns max - one_mutez.