Module Tezos_baking_alpha.Baking_pow

val empty_proof_of_work_nonce : Tezos_base.TzPervasives.Bytes.t

A null proof-of-work nonce. This should only be used to nonsensical blocks of the correct size and shape.

val mine : proof_of_work_threshold:int64 -> Tezos_base.TzPervasives.Block_header.shell_header -> (Tezos_base.TzPervasives.Bytes.t -> Tezos_protocol_alpha.Protocol.Alpha_context.Block_header.contents) -> Tezos_protocol_alpha.Protocol.Alpha_context.Block_header.contents Tezos_base.TzPervasives.tzresult Lwt.t

mine ~proof_of_work_threshold chain block header builder returns a block with a valid proof-of-work nonce. The function builder, provided by the caller, is used to make the block. All the internal logic of generating nonces and checking for the proof-of-work threshold is handled by mine.