Module Tezos_baking_alpha.Operation_pool

module Operation_set : Tezos_base.TzPervasives.Set.S with type elt = Tezos_protocol_alpha.Protocol.Alpha_context.packed_operation

type 'collection t = {

consensus : 'collection;

votes : 'collection;

anonymous : 'collection;

managers : 'collection;

}

Generic base type for pools

type pool = Operation_set.t t

A pool of operations for a single origin, or undifferenciated origin, typically used for operations coming from the node

A pool of operations for a single origin, or undifferenciated origin, typically used for operations coming from the node

val empty : pool

on pool

val pp_pool : Stdlib.Format.formatter -> pool -> unit

val filter_pool : (Tezos_protocol_alpha.Protocol.Alpha_context.packed_operation -> bool) -> pool -> pool

val add_operation : pool -> Tezos_protocol_alpha.Protocol.Alpha_context.packed_operation -> pool

val add_operations : pool -> Tezos_protocol_alpha.Protocol.Alpha_context.packed_operation list -> pool

type ordered_pool = Tezos_protocol_alpha.Protocol.Alpha_context.packed_operation list t

Ordered pool of operations

val ordered_pool_encoding : ordered_pool Tezos_base.TzPervasives.Data_encoding.t

val empty_ordered : ordered_pool

val pp_ordered_pool : Stdlib.Format.formatter -> ordered_pool -> unit

type payload = {

votes_payload : Tezos_protocol_alpha.Protocol.Alpha_context.packed_operation list;

anonymous_payload : Tezos_protocol_alpha.Protocol.Alpha_context.packed_operation list;

managers_payload : Tezos_protocol_alpha.Protocol.Alpha_context.packed_operation list;

}

val empty_payload : payload

val payload_encoding : payload Tezos_base.TzPervasives.Data_encoding.t

val pp_payload : Stdlib.Format.formatter -> payload -> unit

val payload_of_ordered_pool : ordered_pool -> payload

val ordered_pool_of_payload : consensus_operations: Tezos_protocol_alpha.Protocol.Alpha_context.packed_operation list -> payload -> ordered_pool

type consensus_filter = {

level : int32;

round : Tezos_protocol_alpha.Protocol.Alpha_context.Round.t;

payload_hash : Tezos_protocol_alpha.Protocol.Block_payload_hash.t;

}

val filter_with_relevant_consensus_ops : attestation_filter:consensus_filter -> preattestation_filter:consensus_filter option -> Operation_set.t -> Operation_set.t

val unpack_preattestation : Tezos_protocol_alpha.Protocol.Alpha_context.packed_operation -> Tezos_protocol_alpha.Protocol.Alpha_context.Kind.preattestation Tezos_protocol_alpha.Protocol.Alpha_context.operation option

val unpack_attestation : Tezos_protocol_alpha.Protocol.Alpha_context.packed_operation -> Tezos_protocol_alpha.Protocol.Alpha_context.Kind.attestation Tezos_protocol_alpha.Protocol.Alpha_context.operation option

val filter_preattestations : Tezos_protocol_alpha.Protocol.Alpha_context.packed_operation list -> Tezos_protocol_alpha.Protocol.Alpha_context.Kind.preattestation Tezos_protocol_alpha.Protocol.Alpha_context.operation list

val filter_attestations : Tezos_protocol_alpha.Protocol.Alpha_context.packed_operation list -> Tezos_protocol_alpha.Protocol.Alpha_context.Kind.attestation Tezos_protocol_alpha.Protocol.Alpha_context.operation list

val ordered_to_list_list : ordered_pool -> Tezos_protocol_alpha.Protocol.Alpha_context.packed_operation list list

val ordered_of_list_list : Tezos_protocol_alpha.Protocol.Alpha_context.packed_operation list list -> ordered_pool option

val extract_operations_of_list_list : Tezos_protocol_alpha.Protocol.Alpha_context.packed_operation list list -> (Tezos_protocol_alpha.Protocol.Alpha_context.Kind.preattestation Tezos_protocol_alpha.Protocol.Alpha_context.operation list option * Tezos_protocol_alpha.Protocol.Alpha_context.Kind.attestation Tezos_protocol_alpha.Protocol.Alpha_context.operation list * payload) option

preattestation <> None => (List.length preattestations > 0)

module Prioritized_operation : sig ... end

module Prioritized_operation_set : sig ... end

module Prioritized : sig ... end

Pool of prioritized operations