Module Tezos_baking_alpha.Vdf_helpers

val is_in_nonce_revelation_stage : nonce_revelation_threshold:int32 -> level:Tezos_protocol_alpha.Protocol.Alpha_context.Level.t -> bool

is_in_nonce_revelation_stage ~nonce_revelation_threshold ~level checks whether level is part of the nonce revelation stage of its cycle, which is defined as the first nonce_revelation_threshold blocks of every cycle. It is used to avoid calling the Seed_computation RPC on blocks which are part of the nonce revelation stage.