Module Tezos_client_alpha.Injection

type 'kind preapply_result = Tezos_base.TzPervasives.Operation_hash.t * 'kind Tezos_protocol_alpha.Protocol.Alpha_context.operation * 'kind Tezos_protocol_alpha.Protocol.Apply_results.operation_metadata

type fee_parameter = {

minimal_fees : Tezos_protocol_alpha.Protocol.Alpha_context.Tez.t;

minimal_nanotez_per_byte : Q.t;

minimal_nanotez_per_gas_unit : Q.t;

force_low_fee : bool;

fee_cap : Tezos_protocol_alpha.Protocol.Alpha_context.Tez.t;

burn_cap : Tezos_protocol_alpha.Protocol.Alpha_context.Tez.t;

}

val preapply : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> ?verbose_signing:bool -> ?fee_parameter:fee_parameter -> ?branch:int -> ?src_sk:Tezos_client_base.Client_keys.sk_uri -> 'kind Tezos_protocol_alpha.Protocol.Alpha_context.contents_list -> 'kind preapply_result Tezos_base.TzPervasives.tzresult Lwt.t

val simulate : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> ?successor_level:bool -> ?branch:int -> ?latency:int -> 'kind Tezos_protocol_alpha.Protocol.Alpha_context.contents_list -> 'kind preapply_result Tezos_base.TzPervasives.tzresult Lwt.t

Perform simulation of the given operations and return the corresponding preapply_results.

type 'kind result_list = Tezos_base.TzPervasives.Operation_hash.t * 'kind Tezos_protocol_alpha.Protocol.Alpha_context.contents_list * 'kind Tezos_protocol_alpha.Protocol.Apply_results.contents_result_list

val inject_operation : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> ?confirmations:int -> ?dry_run:bool -> ?simulation:bool -> ?successor_level:bool -> ?branch:int -> ?src_sk:Tezos_client_base.Client_keys.sk_uri -> ?verbose_signing:bool -> ?fee_parameter:fee_parameter -> 'kind Tezos_protocol_alpha.Protocol.Alpha_context.contents_list -> 'kind result_list Tezos_base.TzPervasives.tzresult Lwt.t

/!\ inject_operation does not perform automatic patching of gas, storage and fees; use inject_manager_operation to inject manager operations.

type 'kind result = Tezos_base.TzPervasives.Operation_hash.t * 'kind Tezos_protocol_alpha.Protocol.Alpha_context.contents * 'kind Tezos_protocol_alpha.Protocol.Apply_results.contents_result

val prepare_manager_operation : fee:Tezos_protocol_alpha.Protocol.Alpha_context.Tez.t Limit.t -> gas_limit: Tezos_protocol_alpha.Protocol.Alpha_context.Gas.Arith.integral Limit.t -> storage_limit:Z.t Limit.t -> 'kind Tezos_protocol_alpha.Protocol.Alpha_context.manager_operation -> 'kind Annotated_manager_operation.t

val inject_manager_operation : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> ?successor_level:bool -> ?branch:int -> ?confirmations:int -> ?dry_run:bool -> ?verbose_signing:bool -> ?simulation:bool -> ?force:bool -> ?safety_guard:Tezos_protocol_alpha.Protocol.Alpha_context.Gas.Arith.integral -> source:Tezos_base.TzPervasives.Signature.Public_key_hash.t -> src_pk:Tezos_base.TzPervasives.Signature.public_key -> src_sk:Tezos_client_base.Client_keys.sk_uri -> fee:Tezos_protocol_alpha.Protocol.Alpha_context.Tez.t Limit.t -> gas_limit: Tezos_protocol_alpha.Protocol.Alpha_context.Gas.Arith.integral Limit.t -> storage_limit:Z.t Limit.t -> ?counter:Tezos_protocol_alpha.Protocol.Alpha_context.Manager_counter.t -> ?replace_by_fees:bool -> fee_parameter:fee_parameter -> 'kind Annotated_manager_operation.annotated_list -> (Tezos_base.TzPervasives.Operation_hash.t * Tezos_protocol_alpha.Protocol.Alpha_context.packed_operation * 'kind Tezos_protocol_alpha.Protocol.Alpha_context.Kind.manager Tezos_protocol_alpha.Protocol.Alpha_context.contents_list * 'kind Tezos_protocol_alpha.Protocol.Alpha_context.Kind.manager Tezos_protocol_alpha.Protocol.Apply_results.contents_result_list) Tezos_base.TzPervasives.tzresult Lwt.t

val originated_contracts : force:bool -> 'kind Tezos_protocol_alpha.Protocol.Apply_results.contents_result_list -> Tezos_protocol_alpha.Protocol.Contract_hash.t list Tezos_base.TzPervasives.tzresult

Collects the addresses of all contracts originated by a batch of operations by looking at the operation results. Fails if an operation in the batch is failed unless force is given.