Module Tezos_client_alpha.Michelson_v1_printer

val print_expr : Stdlib.Format.formatter -> Tezos_protocol_alpha.Protocol.Script_repr.expr -> unit

val print_expr_unwrapped : Stdlib.Format.formatter -> Tezos_protocol_alpha.Protocol.Script_repr.expr -> unit

val unparse_stack : 'location -> (Tezos_protocol_alpha.Protocol.Script_repr.expr * Tezos_protocol_alpha.Protocol.Script_repr.expr) list -> ('location, string) Tezos_micheline.Micheline.node

val print_typed_stack : Stdlib.Format.formatter -> (Tezos_protocol_alpha.Protocol.Script_repr.expr * Tezos_protocol_alpha.Protocol.Script_repr.expr) list -> unit

val print_execution_trace : Stdlib.Format.formatter -> Tezos_protocol_alpha.Protocol.Script_typed_ir.execution_trace -> unit

val print_big_map_diff : Stdlib.Format.formatter -> Tezos_protocol_alpha.Protocol.Alpha_context.Lazy_storage.diffs -> unit

val inject_types : Tezos_protocol_alpha.Protocol.Script_tc_errors.type_map -> Michelson_v1_parser.parsed -> Tezos_micheline.Micheline_printer.node

Insert the type map returned by the typechecker as comments in a printable Micheline AST.

val unparse_toplevel : ?type_map:Tezos_protocol_alpha.Protocol.Script_tc_errors.type_map -> Tezos_protocol_alpha.Protocol.Alpha_context.Script.expr -> Michelson_v1_parser.parsed

Unexpand the macros and produce the result of parsing an intermediate pretty printed source. Useful when working with contracts extracted from the blockchain and not local files.

val unparse_expression : Tezos_protocol_alpha.Protocol.Alpha_context.Script.expr -> Michelson_v1_parser.parsed

val unparse_invalid : string Tezos_micheline.Micheline.canonical -> string Michelson_v1_parser.parser_result

Unexpand the macros and produce the result of parsing an intermediate pretty printed source. Works on generic trees,for programs that fail to be converted to a specific script version.

val ocaml_constructor_of_prim : Tezos_protocol_alpha.Protocol.Michelson_v1_primitives.prim -> string

val micheline_string_of_expression : zero_loc:bool -> Tezos_protocol_alpha.Protocol.Alpha_context.Script.expr -> string