Module Tezos_client_alpha.Protocol_client_context

module Alpha_block_services : sig ... end

class type rpc_context = object ... end

Client RPC context

class wrap_rpc_context : Tezos_rpc.Context.generic -> rpc_context

The class wrap_rpc_context is a wrapper class used by the proxy mode clients. From a general-purpose Tezos_rpc.Context.generic t, the class is augmented with shell services to provide RPC calls that are protocol-dependent.

class type full = object ... end

The class type full allows to create contexts that are explicitly used by low-level shell functions, while containing various information (I/O services, RPCs...). Then, depending on the usage, the type may be coerced into one of its following ascendants to serve for explicit operations on blocks, chain or daemon for instance.

class wrap_full : Tezos_client_base.Client_context.full -> full

From a Client_context.full, the class allows to call RPCs from the node and those defined by the protocol.

val register_error_kind : Tezos_base.TzPervasives.error_category -> id:string -> title:string -> description:string -> ?pp:(Stdlib.Format.formatter -> 'a -> unit) -> 'a Data_encoding.t -> (Tezos_base.TzPervasives.error -> 'a option) -> ('a -> Tezos_base.TzPervasives.error) -> unit